ALAE-36: Changed duration of Audio to Long

+++ app/src/main/java/com/audiolearning/app/adapters/recycler_view_adapter/AudioRecyclerViewAdapter.kt
@@ -25,7 +25,7 @@ class AudioRecyclerViewAdapter(private var listener: ItemSelectListener) :
         holder.itemView.tv_audio_create_date.text =
             (data[position] as Audio).createDate.toFormattedDate()
         holder.itemView.tv_audio_duration.text = getTimeStringFromMillis(
-            (data[position] as Audio).durationInMilliseconds.toLong()
+            (data[position] as Audio).durationInMilliseconds
         )
     }
